+++ tests/bluetooth/audio/ascs/src/test_ase_state_transition_invalid.c
@@ -375,7 +375,8 @@ static void expect_ase_state_releasing(struct bt_conn *conn, const struct bt_gat
 	zexpect_not_null(ase);
 
 	ret = ase->read(conn, ase, &hdr, sizeof(hdr), 0);
-	zassert_false(ret < 0, "attr->read returned unexpected (err 0x%02x)", BT_GATT_ERR(ret));
+	zassert_false(ret < 0, "attr->read returned unexpected (err 0x%02x)",
+		      (uint8_t)BT_GATT_ERR(ret));
 	zassert_equal(BT_BAP_EP_STATE_RELEASING, hdr.ase_state,
 		      "unexpected ASE_State 0x%02x", hdr.ase_state);
 }
